+++ TOPdeskPS/functions/General API/Send-TdIncidentFile.ps1
@@ -51,9 +51,11 @@
     process {
         $uri = (Get-TdUrl) + "/tas/api/incidents/number/$Number/attachments"
         Write-PSFMessage "Uri - $uri" -Level debug
+        #TODO throw this into an internal function and clean this up a bit.
         $fileName = Split-Path $File -leaf
         $boundary = [guid]::NewGuid().ToString()
         $fileBin = [System.IO.File]::ReadAllBytes($File)
+        $LF = "`r`n"
         $enc = [System.Text.Encoding]::GetEncoding("iso-8859-1")
         $fileEnc = $enc.GetString($fileBin)
 
@@ -78,39 +80,14 @@
             "--$boundary--$LF"
 
         ) -join $LF
-        $Headers = @{
-            'Authorization' = $Script:LoginToken
-        }
+        Write-PSFMessage $bodyLines -Level debug
         $params = @{
             Uri         = $Uri
             Body        = $bodyLines
             Method      = 'Post'
-            Headers     = $headers
             ContentType = "multipart/form-data; boundary=$boundary"
         }
-        Invoke-restmethod @params
-        #Invoke-TdMethod @params
+        Invoke-TdMethod @params
     }
     end {}
-    <#
-    Write-PSFMessage "Bound parameters: $($PSBoundParameters.Keys -join ", ")" -Tag 'debug', 'start', 'param' -Level InternalComment
-    $uri = (Get-TdUrl) + "/tas/api/incidents/number/$Number/attachments"
-    if (-not $Script:__LoginToken) {
-        throw 'no connection to topdesk, try running Connect-TdService'
-    }
-    $headers = @{
-        'Authorization' = $Script:__LoginToken
-    }
-    $body = @{
-        File = $File
-    }
-    $params = @{
-        'uri'       = $uri
-        Method      = 'post'
-        Infile      = $File
-        ContentType = 'multipart/form-data'
-        Headers     = $headers
-    }
-    Invoke-restmethod @params
-    #>
 }